backend/pyproject.toml
[project]
name = "assistant-transport-backend"
version = "0.2.0"
description = "Assistant Transport Backend with direct Anthropic SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "assistant-stream>=0.0.31",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.39.0",
    "httpx>=0.24.0",
    "beautifulsoup4>=4.12.0",
    "scrapingbee>=2.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
exclude = ["test_client.py"]

[tool.ruff]
target-version = "py39"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100
